+++ video-wallpaper/settings/AutomationTab.qml
@@ -1,8 +1,10 @@
 pragma ComponentBehavior: Bound
 import QtQuick
+import QtQuick.Controls
 import QtQuick.Layouts
 
 import qs.Commons
+import qs.Services.UI
 import qs.Widgets
 
 ColumnLayout {
@@ -17,9 +19,10 @@ ColumnLayout {
     required property var pluginApi
     required property bool enabled
 
-    property bool   automation:     pluginApi?.pluginSettings?.automation     || false
-    property string automationMode: pluginApi?.pluginSettings?.automationMode || pluginApi?.manifest?.metadata?.defaultSettings?.automationMode || ""
-    property real   automationTime: pluginApi?.pluginSettings?.automationTime || pluginApi?.manifest?.metadata?.defaultSettings?.automationTime || 0
+    property bool      automation:           pluginApi?.pluginSettings?.automation           || false
+    property list<int> automationCustomTime: pluginApi?.pluginSettings?.automationCustomTime || []
+    property string    automationMode:       pluginApi?.pluginSettings?.automationMode       || pluginApi?.manifest?.metadata?.defaultSettings?.automationMode || ""
+    property real      automationTime:       pluginApi?.pluginSettings?.automationTime       || pluginApi?.manifest?.metadata?.defaultSettings?.automationTime || 0
 
 
     /***************************
@@ -64,45 +67,226 @@ ColumnLayout {
         }
 
         RowLayout {
+            id: timesLayout
             spacing: Style.marginS
-
 
             Repeater {
                 model: [
-                    {
-                        "text": root.pluginApi?.tr("settings.automation.time.5m") || "5m",
-                        "time": 5 * 60
-                    },
-                    {
-                        "text": root.pluginApi?.tr("settings.automation.time.10m") || "10m",
-                        "time": 10 * 60
-                    },
-                    {
-                        "text": root.pluginApi?.tr("settings.automation.time.30m") || "30m",
-                        "time": 30 * 60
-                    },
-                    {
-                        "text": root.pluginApi?.tr("settings.automation.time.1h") || "1h",
-                        "time": 60 * 60
-                    },
-                    {
-                        "text": root.pluginApi?.tr("settings.automation.time.1h30m") || "1h 30m",
-                        "time": 90 * 60
-                    },
-                    {
-                        "text": root.pluginApi?.tr("settings.automation.time.2h") || "2h",
-                        "time": 120 * 60
-                    },
+                    5 * 60,
+                    30 * 60,
+                    60 * 60,
+                    90 * 60,
+                    120 * 60
                 ]
 
-                NButton {
+                TimeButton {
                     required property var modelData
-                    enabled: root.enabled && root.automation
-
-                    text: modelData.text
-                    onClicked: root.automationTime = modelData.time
-                    outlined: root.automationTime === modelData.time
+                    time: modelData
                 }
+            }
+
+            NDivider {
+                vertical: true
+            }
+
+            Repeater {
+                id: customTimeRepeater
+                model: root.automationCustomTime
+
+                CustomTimeButton {
+                    required property var modelData
+                    time: modelData
+                }
+            }
+
+            NIconButton {
+                icon: "plus"
+                tooltipText: root.pluginApi?.tr("settings.automation.time.custom.create.tooltip") || "Create a custom time."
+                enabled: root.enabled && root.automation
+                onClicked: createCustomTime.open();
+            }
+
+            component TimeButton: NButton {
+                required property int time
+
+                text: {
+                    const hour = Math.floor(time / 3600.0);
+                    const minute = Math.max(Math.floor(time / 60.0), 1) - (hour * 60);
+                    const hourTranslation = root.pluginApi?.tr("settings.automation.time.h", {hour: hour}) || `${hour}h`;
+                    const minuteTranslation = root.pluginApi?.tr("settings.automation.time.m", {minute: minute}) || `${minute}m`;
+
+                    if (hour == 0) {
+                        return minuteTranslation;
+                    } else if (minute == 0) {
+                        return hourTranslation;
+                    } else {
+                        return `${hourTranslation} ${minuteTranslation}`;
+                    }
+                }
+                tooltipText: {
+                    const hour = Math.floor(time / 3600.0);
+                    const minute = Math.max(Math.floor(time / 60.0), 1) - (hour * 60);
+                    const hourTranslation = root.pluginApi?.trp("settings.automation.time.hour", hour, "1 hour", "{count} hours") || "";
+                    const minuteTranslation = root.pluginApi?.trp("settings.automation.time.minute", minute, "1 minute", "{count} minutes") || "";
+
+                    if (hour == 0) {
+                        return minuteTranslation;
+                    } else if (minute == 0) {
+                        return hourTranslation;
+                    } else {
+                        return `${hourTranslation} ${minuteTranslation}`;
+                    }
+                }
+                enabled: root.enabled && root.automation
+                onClicked: root.automationTime = time
+                outlined: root.automationTime === time
+            }
+
+            component CustomTimeButton: TimeButton {
+                id: timeButton
+
+                onRightClicked: {
+                    customTimeButtonMenu.selectedSeconds = time;
+                    customTimeButtonMenu.openAtItem(timeButton, 0, timeButton.height);
+                }
+            }
+        }
+    }
+
+    Popup {
+        id: createCustomTime
+
+        width: 400
+
+        anchors.centerIn: Overlay.overlay
+        padding: Style.marginL
+
+        modal: true
+        closePolicy: Popup.CloseOnEscape | Popup.CloseOnPressOutside
+
+        background: Rectangle {
+            color: Color.mSurfaceVariant
+            radius: Style.iRadiusL
+            border.color: Color.mOutline
+            border.width: Style.borderS
+        }
+
+        function save() {
+            // Check format of the text
+            if (customTimeInput.isCorrectFormat()) {
+                const times = customTimeInput.text.split(":");
+                const hour = parseInt(times[0]);
+                let minute = parseInt(times[1]);
+
+                if (hour == 0) {
+                    // Make sure that minute isn't 0
+                    minute = Math.max(minute, 1);
+                }
+
+                const seconds = hour * 3600 + minute * 60;
+                root.automationCustomTime.push(seconds);
+
+                // Make sure no duplicates exist.
+                let s = new Set(root.automationCustomTime);
+                root.automationCustomTime = [...s];
+                root.automationCustomTime.sort();
+
+                // Reset the input
+                customTimeInput.text = "";
+
+                // Save the custom time to settings
+                if (root.pluginApi == null) {
+                    Logger.e("video-wallpaper", "Plugin API is null.");
+                    return;
+                }
+                root.pluginApi.pluginSettings.automationCustomTime = root.automationCustomTime;
+                root.pluginApi.saveSettings();
+
+                createCustomTime.close();
+            }
+        }
+
+        ColumnLayout {
+            anchors.fill: parent
+            spacing: Style.marginS
+
+            NTextInput {
+                id: customTimeInput
+                label: root.pluginApi?.tr("settings.automation.time.custom.create.label") || "Add Custom Time"
+                description: root.pluginApi?.tr("settings.automation.time.custom.create.description") || "Enter time as HH:MM, (e.g. 1:20)"
+                placeholderText: root.pluginApi?.tr("settings.automation.time.custom.create.placeholder") || "Example, 1:15"
+                Layout.fillWidth: true
+                onEditingFinished: createCustomTime.save();
+
+                function isCorrectFormat() {
+                    // Check that : is included and that only one : is included
+                    const times = customTimeInput.text.split(":");
+                    if (times.length != 2) {
+                        return false;
+                    }
+
+                    const isInteger = /^[0-9]+$/;
+                    const isMinute = /^[0-5][0-9]$/;
+
+                    // Check so that both sides are numbers and that the minute is in the range of [0-59]
+                    if (!isInteger.test(times[0]) || !isInteger.test(times[1]) || !isMinute.test(times[1])) {
+                        return false;
+                    }
+
+                    return true;
+                }
+            }
+
+            RowLayout {
+                spacing: Style.marginS
+                layoutDirection: Qt.LeftToRight
+                Layout.fillWidth: true
+
+                Item {
+                    Layout.fillWidth: true
+                }
+
+                NButton {
+                    text: root.pluginApi?.tr("settings.automation.time.custom.create.cancel") || "Cancel"
+                    onClicked: createCustomTime.close();
+                }
+
+                NButton {
+                    text: root.pluginApi?.tr("settings.automation.time.custom.create.save") || "Save"
+                    enabled: customTimeInput.isCorrectFormat();
+                    onClicked: createCustomTime.save();
+                }
+            }
+        }
+    }
+
+    NContextMenu {
+        id: customTimeButtonMenu
+        property int selectedSeconds: 0
+
+        model: [
+            {
+                "label": root.pluginApi?.tr("settings.automation.time.custom.remove") || "Remove",
+                "action": "remove",
+                "icon": "x"
+            },
+        ]
+
+        onTriggered: action => {
+            close();
+
+            switch(action) {
+                case "remove":
+                    const index = root.automationCustomTime.indexOf(selectedSeconds);
+                    if(index !== -1) {
+                        // Remove the element from the list
+                        root.automationCustomTime.splice(index, 1);
+                        root.pluginApi.pluginSettings.automationCustomTime = root.automationCustomTime;
+                        root.pluginApi.saveSettings();
+                    }
+                    break;
+                default:
+                    Logger.e("video-wallpaper", "Error, action not found:", action);
             }
         }
     }
@@ -111,9 +295,10 @@ ColumnLayout {
         target: root.pluginApi
         function onPluginSettingsChanged() {
             // Update the local properties on change
-            root.automation =     root.pluginApi?.pluginSettings?.automation     || false
-            root.automationMode = root.pluginApi?.pluginSettings?.automationMode || root.pluginApi?.manifest?.metadata?.defaultSettings?.automationMode || ""
-            root.automationTime = root.pluginApi?.pluginSettings?.automationTime || root.pluginApi?.manifest?.metadata?.defaultSettings?.automationTime || 0
+            root.automation =           root.pluginApi?.pluginSettings?.automation           || false
+            root.automationCustomTime = root.pluginApi?.pluginSettings?.automationCustomTime || []
+            root.automationMode =       root.pluginApi?.pluginSettings?.automationMode       || root.pluginApi?.manifest?.metadata?.defaultSettings?.automationMode || ""
+            root.automationTime =       root.pluginApi?.pluginSettings?.automationTime       || root.pluginApi?.manifest?.metadata?.defaultSettings?.automationTime || 0
         }
     }
 
@@ -128,6 +313,7 @@ ColumnLayout {
         }
 
         pluginApi.pluginSettings.automation = automation;
+        pluginApi.pluginSettings.automationCustomTime = automationCustomTime;
         pluginApi.pluginSettings.automationMode = automationMode;
         pluginApi.pluginSettings.automationTime = automationTime;
     }